# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koperasi Sekar Kartini API",
        "description": "Documentation API Koperasi Sekar Kartini",
        "contact": {
            "name": "API Support",
            "email": "support@koperasisekarkartini.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get all employees",
                "operationId": "7fc06f144c1cd7925c5c391a6609f02d",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create new employee",
                "operationId": "bfad4edec29b0edbd684a0afbd847510",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "member_number",
                                    "identity_number",
                                    "birth_date",
                                    "phone_number",
                                    "address",
                                    "occupation",
                                    "identity_card_photo",
                                    "self_photo",
                                    "password",
                                    "work_area_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Siti Nurhaliza"
                                    },
                                    "member_number": {
                                        "description": "Nomor anggota unik",
                                        "type": "string",
                                        "example": "EMP-001"
                                    },
                                    "identity_number": {
                                        "description": "NIK unik",
                                        "type": "string",
                                        "example": "3201010101900002"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-02-15"
                                    },
                                    "phone_number": {
                                        "description": "Nomor telepon unik",
                                        "type": "string",
                                        "example": "081234567891"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 456"
                                    },
                                    "occupation": {
                                        "type": "string",
                                        "example": "Karyawan"
                                    },
                                    "identity_card_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "self_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "work_area_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Employee created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Group": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "number": {
                        "type": "string",
                        "example": "GRP001"
                    },
                    "description": {
                        "type": "string",
                        "example": "Kelompok tani sejahtera",
                        "nullable": true
                    },
                    "shared_liability_fund_amount": {
                        "type": "integer",
                        "example": 50000
                    },
                    "group_fund_amount": {
                        "type": "integer",
                        "example": 30000
                    },
                    "social_fund_amount": {
                        "type": "integer",
                        "example": 20000
                    },
                    "total_shared_liability_fund": {
                        "type": "integer",
                        "example": 500000
                    },
                    "total_group_fund": {
                        "type": "integer",
                        "example": 300000
                    },
                    "total_social_fund": {
                        "type": "integer",
                        "example": 200000
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "work_area_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "chairman_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "facilitator_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "treasurer_id": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Rahmat Admin"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "081200000001"
                    },
                    "member_number": {
                        "type": "string",
                        "example": "ADM-001"
                    },
                    "identity_number": {
                        "type": "string",
                        "example": "3201111111110001"
                    },
                    "address": {
                        "type": "string",
                        "example": "Jl. Anggur No. 10, Sumbersari, Jember"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "is_verified": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "work_area": {
                        "$ref": "#/components/schemas/WorkArea"
                    },
                    "group": {
                        "$ref": "#/components/schemas/Group"
                    }
                },
                "type": "object"
            },
            "WorkArea": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name_work_area": {
                        "type": "string",
                        "example": "Wilayah Kerja A"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use your API token as Bearer token for authorization",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}